% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dungfauna_event}
\alias{dungfauna_event}
\title{A dataset of deliberately introduced dung beetle records (wide format)}
\format{
A data frame with 64 variables and 10,113 rows:
\describe{
\item{\verb{Bubas bison}, \verb{Copris elphenor}, \verb{Copris hispanus},
\verb{Digitonthophagus gazella}, \verb{Euoniticellus africanus},
\verb{Euoniticellus fulvus}, \verb{Euoniticellus intermedius},
\verb{Euoniticellus pallipes}, \verb{Geotrupes spiniger}, \verb{Liatongus militaris},
\verb{Onitis alexis}, \verb{Onitis aygulus}, \verb{Onitis caffer}, \verb{Onitis pecuarius},
\verb{Onitis vanderkelleni}, \verb{Onitis viridulus}, \verb{Onthophagus binodis},
\verb{Onthophagus nigriventris}, \verb{Onthophagus obliquus},
\verb{Onthophagus sagittarius}, \verb{Onthophagus taurus}, \verb{Sisyphus rubrus},
\verb{Sisyphus spinipes}}{Either the number of individual specimens collected,
or the activity level observed, for the given species during the sampling
event. See \strong{Details}}
}
}
\usage{
dungfauna_event
}
\description{
These data provide counts of deliberately introduced dung beetles in
Australia for 10,113 sampling events with each row representing one
sampling event. Column headings use the Darwin Core terms, except for columns
showing species counts (columns from "Bubas bison" to "Sisyphus spinipes")
and column headings that include an underscore (_). In the latter case, the
underscore has been included to indicate that a non Darwin Core term has been
used. The variables are as documented in \code{\link{dungfauna_occurrence}} except for
the variables "Bubas bison" to "Sisyphus spinipes". See below under
\strong{Format}.
}
\details{
The values in the columns "Bubas bison" to "Sisyphus spinipes"
represent the number of individuals of the species collected by the method
described in the column samplingProtocol, except where the value in
eventRemarks is "Species counts are visual activity ratings". In the latter
case the values range from 0 to 5, with 0 indicating the species was not
found, and 5 indicating that the species was found in almost every dung pad
searched.
}
\keyword{datasets}
