% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dungfauna_occurrence}
\alias{dungfauna_occurrence}
\title{A dataset of deliberately introduced dung beetle records (long format)}
\format{
A data frame with 58 variables and 232,599 rows
}
\usage{
dungfauna_occurrence
}
\description{
These data provide counts of deliberately introduced dung beetles in
Australia for 23 species and 10,113 sampling events with each row
representing the data for one species from one sampling event. Column
headings use the Darwin Core terms, except for column headings that include
an underscore (_). In the latter case, the underscore has been included to
indicate that a non Darwin Core term has been used.
}
\details{
The data are fully documented in "link to data paper here"
}
\keyword{datasets}
