% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run the shiny app visualisation}
\usage{
runShinyApp()
}
\description{
Starts a visualisation of the dataset in a shiny app that can be
viewed locally in the browser. After calling this function,
If the app window doesn't start automatically,
navigate to the port printed to your console
(usually \url{localhost:5197}) in your web browser.
}
\examples{
# start the visualisation with:
# dungfaunaR::runShinyApp()
}
