% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_stateProvince_aus}
\alias{county_stateProvince_aus}
\title{Shapefile for linking geographic points to country - county Darwin Core terms}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 547 rows and 5 columns.
}
\usage{
county_stateProvince_aus
}
\description{
Within Australia, the next smaller administrative unit after country is
State / Territory, after which comes a 'Local Government Area'. These are
equivalent to the Darwin Core terms stateProvince and county.
}
\details{
The Australian Bureau of Statistics (ABS) provides shapefile data for local
government areas in Australia which also includes the associated higher
level geographies (State / Territory and Country). More information is here:
\url{https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files}.

The download link for the raw data is (Local Government Areas - 2022 -
Shapefile in GDA2020):
https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files/LGA_2022_AUST_GDA2020_SHP.zip
}
\keyword{datasets}
